requirements.txt
django==3.2
djangorestframework==3.13.1
django-filter==22.1
sentry-sdk==1.5.12
django-oauth-toolkit==2.1.0